Update rust edition to 2024

+++ graphannis/src/annis/db/corpusstorage.rs
@@ -15,28 +15,28 @@ use crate::annis::types::{
     QueryAttributeDescription,
 };
 use crate::annis::types::{CorpusSizeInfo, CountExtra};
-use crate::annis::util::quicksort;
 use crate::annis::util::TimeoutCheck;
-use crate::{graph::Match, AnnotationGraph};
+use crate::annis::util::quicksort;
+use crate::{AnnotationGraph, graph::Match};
 use fmt::Display;
 use fs2::FileExt;
 use graphannis_core::annostorage::symboltable::SymbolTable;
 use graphannis_core::annostorage::{
-    match_group_resolve_symbol_ids, match_group_with_symbol_ids, NodeAnnotationStorage,
+    NodeAnnotationStorage, match_group_resolve_symbol_ids, match_group_with_symbol_ids,
 };
 use graphannis_core::errors::Result as CoreResult;
 use graphannis_core::{
     annostorage::{MatchGroup, ValueSearch},
     graph::{
-        storage::GraphStatistic, update::GraphUpdate, ANNIS_NS, NODE_NAME, NODE_NAME_KEY, NODE_TYPE,
+        ANNIS_NS, NODE_NAME, NODE_NAME_KEY, NODE_TYPE, storage::GraphStatistic, update::GraphUpdate,
     },
     types::{AnnoKey, Annotation, Component, NodeID},
 };
 use itertools::Itertools;
 use linked_hash_map::LinkedHashMap;
 use memory_stats::memory_stats;
-use percent_encoding::{percent_decode_str, utf8_percent_encode, AsciiSet, CONTROLS};
-use rand::Rng;
+use percent_encoding::{AsciiSet, CONTROLS, percent_decode_str, utf8_percent_encode};
+use rand::prelude::*;
 use smartstring::alias::String as SmartString;
 use std::collections::HashSet;
 use std::fmt;
@@ -53,7 +53,6 @@ use transient_btree_index::{BtreeConfig, BtreeIndex};
 
 use rustc_hash::{FxHashMap, FxHashSet};
 
-use rand::seq::SliceRandom;
 use std::{
     ffi::CString,
     io::{BufReader, Write},
@@ -550,7 +549,7 @@ impl CorpusStorage {
             .unwrap_or_default();
 
         let corpus_info: CorpusInfo = match &*lock {
-            CacheEntry::Loaded(ref db) => {
+            CacheEntry::Loaded(db) => {
                 // check if all components are loaded
                 let mut load_status = LoadStatus::FullyLoaded;
 
@@ -806,10 +805,10 @@ impl CorpusStorage {
                         if let Some(relannis_root) = output_path.parent() {
                             relannis_files.push(relannis_root.to_owned())
                         }
-                    } else if let Some(ext) = output_path.extension() {
-                        if ext.to_string_lossy().to_ascii_lowercase() == "graphml" {
-                            graphannis_files.push(output_path.clone());
-                        }
+                    } else if let Some(ext) = output_path.extension()
+                        && ext.to_string_lossy().to_ascii_lowercase() == "graphml"
+                    {
+                        graphannis_files.push(output_path.clone());
                     }
                 }
 
@@ -1001,10 +1000,10 @@ impl CorpusStorage {
                 .map(|db_entry| db_entry.write())
                 .transpose()?;
 
-            if db_path.is_dir() {
-                if let Err(e) = std::fs::remove_dir_all(&db_path) {
-                    error!("Error when removing existing files {}", e);
-                }
+            if db_path.is_dir()
+                && let Err(e) = std::fs::remove_dir_all(&db_path)
+            {
+                error!("Error when removing existing files {}", e);
             }
         } else if cache.contains_key(&corpus_name) || db_path.is_dir() {
             return Err(GraphAnnisError::CorpusExists(corpus_name.to_string()));
@@ -1084,30 +1083,30 @@ impl CorpusStorage {
             // Get the linked file for this node
             if let Some(original_path) = node_annos.get_value_for_item(&node, &linked_file_key)? {
                 let original_path = old_base_path.join(PathBuf::from(original_path.as_ref()));
-                if original_path.is_file() {
-                    if let Some(node_name) = node_annos.get_value_for_item(&node, &NODE_NAME_KEY)? {
-                        // Create a new file name based on the node name and copy the file
-                        let new_path = new_base_path.join(node_name.as_ref());
-                        debug!(
-                            "Copying file from {} to {}",
-                            original_path.as_path().to_string_lossy(),
-                            new_path.to_string_lossy()
-                        );
-                        if let Some(parent) = new_path.parent() {
-                            std::fs::create_dir_all(parent)?;
-                        }
-                        std::fs::copy(&original_path, &new_path)?;
-                        // Update the annotation to link to the new file with a relative path.
-                        // Use the corpus directory as base path for this relative path.
-                        let relative_path = new_path.strip_prefix(new_base_path)?;
-                        node_annos.insert(
-                            node,
-                            Annotation {
-                                key: linked_file_key.clone(),
-                                val: relative_path.to_string_lossy().into(),
-                            },
-                        )?;
+                if original_path.is_file()
+                    && let Some(node_name) = node_annos.get_value_for_item(&node, &NODE_NAME_KEY)?
+                {
+                    // Create a new file name based on the node name and copy the file
+                    let new_path = new_base_path.join(node_name.as_ref());
+                    debug!(
+                        "Copying file from {} to {}",
+                        original_path.as_path().to_string_lossy(),
+                        new_path.to_string_lossy()
+                    );
+                    if let Some(parent) = new_path.parent() {
+                        std::fs::create_dir_all(parent)?;
                     }
+                    std::fs::copy(&original_path, &new_path)?;
+                    // Update the annotation to link to the new file with a relative path.
+                    // Use the corpus directory as base path for this relative path.
+                    let relative_path = new_path.strip_prefix(new_base_path)?;
+                    node_annos.insert(
+                        node,
+                        Annotation {
+                            key: linked_file_key.clone(),
+                            val: relative_path.to_string_lossy().into(),
+                        },
+                    )?;
                 }
             }
         }
@@ -1719,10 +1718,9 @@ impl CorpusStorage {
                 if let Some(v) = db
                     .get_node_annos()
                     .get_value_for_item(&m.node, &m.anno_key)?
+                    && v == "3.3"
                 {
-                    if v == "3.3" {
-                        relannis_version_33 = true;
-                    }
+                    relannis_version_33 = true;
                 }
             }
         }
@@ -1747,16 +1745,16 @@ impl CorpusStorage {
 
             if find_arguments.order == ResultOrder::Randomized {
                 // Use a unique random index for each match to force a random order
-                let mut rng = rand::thread_rng();
+                let mut rng = rand::rng();
 
                 for mgroup in plan {
                     let mgroup = mgroup?;
-                    let mut idx: usize = rng.gen();
-                    while tmp_results.contains_key(&idx)? {
-                        idx = rng.gen();
+                    let mut idx: u64 = rng.random();
+                    while tmp_results.contains_key(&(idx as usize))? {
+                        idx = rng.random();
                     }
                     let m = match_group_with_symbol_ids(&mgroup, &mut anno_key_symbols)?;
-                    tmp_results.insert(idx, m)?;
+                    tmp_results.insert(idx as usize, m)?;
                 }
             } else {
                 // Insert results in the order as they are given by the iterator
@@ -2022,7 +2020,7 @@ impl CorpusStorage {
                 if order == ResultOrder::Randomized {
                     // This is still oddly ordered, because results from one corpus will always be grouped together.
                     // But it still better than just output the same corpus first.
-                    let mut rng = rand::thread_rng();
+                    let mut rng = rand::rng();
                     corpus_names.shuffle(&mut rng);
                 } else if order == ResultOrder::Inverted {
                     corpus_names.sort();
@@ -2052,11 +2050,11 @@ impl CorpusStorage {
                         result.extend(single_result.into_iter());
                     }
 
-                    if let Some(limit) = limit {
-                        if result.len() == limit {
-                            // Searching in the first corpora already yielded enough results
-                            break;
-                        }
+                    if let Some(limit) = limit
+                        && result.len() == limit
+                    {
+                        // Searching in the first corpora already yielded enough results
+                        break;
                     }
                     if skipped < offset {
                         find_arguments.offset -= skipped;
@@ -2492,36 +2490,36 @@ impl CorpusStorage {
             self.get_loaded_entry_with_components(corpus_name, vec![component.clone()])
         {
             let lock = db_entry.read()?;
-            if let Ok(db) = get_read_or_error(&lock) {
-                if let Some(gs) = db.get_graphstorage(component) {
-                    let edge_annos = gs.get_anno_storage();
-                    for key in edge_annos.annotation_keys()? {
-                        if list_values {
-                            if only_most_frequent_values {
-                                // get the first value
-                                if let Some(val) =
-                                    edge_annos.get_all_values(&key, true)?.into_iter().next()
-                                {
-                                    result.push(Annotation {
-                                        key: key.clone(),
-                                        val: val.into(),
-                                    });
-                                }
-                            } else {
-                                // get all values
-                                for val in edge_annos.get_all_values(&key, false)? {
-                                    result.push(Annotation {
-                                        key: key.clone(),
-                                        val: val.into(),
-                                    });
-                                }
+            if let Ok(db) = get_read_or_error(&lock)
+                && let Some(gs) = db.get_graphstorage(component)
+            {
+                let edge_annos = gs.get_anno_storage();
+                for key in edge_annos.annotation_keys()? {
+                    if list_values {
+                        if only_most_frequent_values {
+                            // get the first value
+                            if let Some(val) =
+                                edge_annos.get_all_values(&key, true)?.into_iter().next()
+                            {
+                                result.push(Annotation {
+                                    key: key.clone(),
+                                    val: val.into(),
+                                });
                             }
                         } else {
-                            result.push(Annotation {
-                                key: key.clone(),
-                                val: SmartString::new(),
-                            });
+                            // get all values
+                            for val in edge_annos.get_all_values(&key, false)? {
+                                result.push(Annotation {
+                                    key: key.clone(),
+                                    val: val.into(),
+                                });
+                            }
                         }
+                    } else {
+                        result.push(Annotation {
+                            key: key.clone(),
+                            val: SmartString::new(),
+                        });
                     }
                 }
             }
@@ -2585,17 +2583,18 @@ impl Drop for CorpusStorage {
             // administration account (see
             // https://github.com/korpling/graphANNIS/issues/230).
             let lock_file_path = self.db_dir.join(DB_LOCK_FILE_NAME);
-            if lock_file_path.exists() && lock_file_path.is_file() {
-                if let Err(e) = std::fs::remove_file(lock_file_path) {
-                    warn!("Could not remove CorpusStorage lock file: {:?}", e);
-                }
+            if lock_file_path.exists()
+                && lock_file_path.is_file()
+                && let Err(e) = std::fs::remove_file(lock_file_path)
+            {
+                warn!("Could not remove CorpusStorage lock file: {:?}", e);
             }
         }
     }
 }
 
 fn get_read_or_error<'a>(lock: &'a RwLockReadGuard<CacheEntry>) -> Result<&'a AnnotationGraph> {
-    if let CacheEntry::Loaded(ref db) = &**lock {
+    if let CacheEntry::Loaded(db) = &**lock {
         Ok(db)
     } else {
         Err(GraphAnnisError::LoadingGraphFailed {
@@ -2607,7 +2606,7 @@ fn get_read_or_error<'a>(lock: &'a RwLockReadGuard<CacheEntry>) -> Result<&'a An
 fn get_write_or_error<'a>(
     lock: &'a mut RwLockWriteGuard<CacheEntry>,
 ) -> Result<&'a mut AnnotationGraph> {
-    if let CacheEntry::Loaded(ref mut db) = &mut **lock {
+    if let CacheEntry::Loaded(db) = &mut **lock {
         Ok(db)
     } else {
         Err(CorpusStorageError::CorpusCacheEntryNotLoaded.into())
@@ -2622,8 +2621,8 @@ fn get_max_cache_size(cache_strategy: &CacheStrategy, used_cache_size: usize) ->
             if let Ok(mem) = sys_info::mem_info() {
                 // the free memory
                 let free_system_mem: usize = mem.avail as usize * 1024; // mem.free is in KiB
-                                                                        // A part of the system memory is already used by the cache.
-                                                                        // We want x percent of the overall available memory (thus not used by us), so add the cache size
+                // A part of the system memory is already used by the cache.
+                // We want x percent of the overall available memory (thus not used by us), so add the cache size
                 let available_memory: usize = free_system_mem + used_cache_size;
                 ((available_memory as f64) * (max_percent / 100.0)) as usize
             } else {
